Move side content to separate namespace
