+++ src/i18n/locales/zh-SG/sideContent/substVisualizer.json
@@ -1,0 +1,13 @@
+{
+  "welcome": "欢迎来到步骤器！",
+  "instructions": "在此标签页中，REPL 将被隐藏，因此请确保您的代码在运行步骤器之前没有错误。您可以通过在左侧编写程序，然后拖动上方的滑块来查看其评估过程。",
+  "evaluationSteps": "在偶数步骤中，将要评估的程序部分会以黄色突出显示。在奇数步骤中，评估结果会以绿色突出显示。您可以在控制栏中更改最大步骤限制（500-5000，默认值为1000）。",
+  "shortcutsTitle": "一些有用的键盘快捷键：",
+  "shortcuts": {
+    "a": "a: 跳到第一步",
+    "e": "e: 跳到最后一步",
+    "f": "f: 跳到下一步",
+    "b": "b: 跳到上一步"
+  },
+  "shortcutsNote": "请注意，这些快捷键仅在浏览器焦点位于此标签页时有效（点击说明文字或其上方区域）。"
+}
